いいねされた追加　いいねは他人のみ　post/showにユーザー画像追加

+++ app/controllers/likes_controller.rb
@@ -4,12 +4,16 @@ class LikesController < ApplicationController
   def create
     @post = Post.find(params[:post_id])
     like = current_user.likes.build(post_id: params[:post_id])
-    like.save
+    unless @post.user == current.user
+      like.save
+    end
   end
 
   def destroy
     @post = Post.find(params[:post_id])
     like = Like.find_by(post_id: params[:post_id], user_id: current_user.id)
-    like.destroy
+    unless @post.user == current.user
+      like.destroy
+    end
   end
 end
